+++ rage/src/bin/rage/error.rs
@@ -29,7 +29,6 @@ pub(crate) enum EncryptError {
     },
     IdentityRead(age::cli_common::ReadError),
     Io(io::Error),
-    MissingRecipients,
     MixedIdentityAndPassphrase,
     MixedRecipientAndPassphrase,
     MixedRecipientsFileAndPassphrase,
@@ -63,6 +62,10 @@ impl From<io::Error> for EncryptError {
 impl fmt::Display for EncryptError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            EncryptError::Age(e @ age::EncryptError::MissingRecipients) => {
+                writeln!(f, "{}", e)?;
+                wfl!(f, "rec-enc-missing-recipients")
+            }
             EncryptError::Age(e) => write!(f, "{}", e),
             EncryptError::BrokenPipe { is_stdout, source } => {
                 if *is_stdout {
@@ -74,10 +77,6 @@ impl fmt::Display for EncryptError {
             }
             EncryptError::IdentityRead(e) => write!(f, "{}", e),
             EncryptError::Io(e) => write!(f, "{}", e),
-            EncryptError::MissingRecipients => {
-                wlnfl!(f, "err-enc-missing-recipients")?;
-                wfl!(f, "rec-enc-missing-recipients")
-            }
             EncryptError::MixedIdentityAndPassphrase => {
                 wfl!(f, "err-enc-mixed-identity-passphrase")
             }